% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_core_variables}
\alias{get_core_variables}
\title{Get core variables required for analysis}
\usage{
get_core_variables(raw_data_clean)
}
\arguments{
\item{raw_data_clean}{A roughly cleaned/processed raw dataset}
}
\value{
A data.frame to which basic tidying and cleaning has been done in
preparation for further data procesing
}
\description{
Get core variables required for analysis
}
\examples{
  \dontrun{
    get_core_variables(raw_data_clean)
  }

}
