% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_na_type}
\alias{get_na_type}
\title{Get NA types}
\usage{
get_na_type(x)
}
\arguments{
\item{x}{A vector of values to determine NA type to use}
}
\value{
An NA type specific for the class type of \code{x}
}
\description{
Determine the NA type needed for recoding a vector of values
}
\examples{
  get_na_type(x = 1:3)
  get_na_type(x = c("1", "2", "3"))

}
