% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesquisas-package.R
\docType{package}
\name{perquisas}
\alias{pesquisas}
\alias{pesquisas-package}
\title{Utilities to Perform Data Cleaning, Processing, and Analysis on Mozambique
Health and Nutrition Surveys}
\description{
In 2019, UNICEF with the support of the European Union launched a 4-year
nutrition programme aimed at contributing to recovery efforts for people in
need in cyclone-affected districts of Mozambique. Since then, a series of
surveys have been conducted by UNICEF and its partners in Mozambique to
provide the data needed to monitor progress and assess change in health and
nutrition outcomes. This package provides R utilities and tools for
performing data cleaning, data processing, and data analysis of these
surveys.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://katilingban.io/pesquisas/}
  \item \url{https://github.com/katilingban/pesquisas}
  \item Report bugs at \url{https://github.com/katilingban/pesquisas/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
