% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{recode_yes_no}
\alias{recode_yes_no}
\title{Recode yes no responses}
\usage{
recode_yes_no(x, na_values = NULL, detect = c("yes", "no"))
}
\arguments{
\item{x}{A vector to recode}

\item{na_values}{A vector of values in x that correspond to an NA. Default
to NULL}

\item{detect}{Whether to recode to an affirmative response ("yes") or to
a negative response ("no")}
}
\value{
A vector of the same length as \code{x} of recoded values
}
\description{
Recode yes no responses
}
\examples{

  recode_yes_no(c(1, 2, 1, 2))
  recode_yes_no(c("yes", "no", "yes", "no"))
  recode_yes_no(c("Yes", "No", "YES", "NO"))

}
