% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{spread_vector_to_columns}
\alias{spread_vector_to_columns}
\title{Convert character vector of categorical responses into unique variables}
\usage{
spread_vector_to_columns(x, fill = NULL, na_rm = FALSE, prefix)
}
\arguments{
\item{x}{Vector of categorical values}

\item{fill}{A vector of all the possible responses that can be selected in
select multiple}

\item{na_rm}{Logical. Should NA responses be made into its own colum? Default
to FALSE}

\item{prefix}{A character string to prepend to the names of the new columns
to be created}
}
\value{
A data.frame with columns for each type of possible responses
(if fill is specified) or for all responses provided.
}
\description{
Function transforms a vector of categorical responses into \code{n} number of
new columns/variables equal to the number of unique categorical values.
}
\examples{
  spread_vector_to_columns(x = 1:4, prefix = "test")

}
